+++ crates/sourcepawn-studio/src/line_index.rs
@@ -9,6 +9,8 @@
 
 use std::sync::Arc;
 
+use rowan::{TextRange, TextSize};
+
 #[derive(Clone, Copy)]
 pub enum PositionEncoding {
     Utf8,
@@ -20,6 +22,42 @@ pub(crate) struct LineIndex {
     #[allow(unused)]
     pub(crate) endings: LineEndings,
     pub(crate) encoding: PositionEncoding,
+}
+
+impl LineIndex {
+    pub fn range(&self, range: TextRange) -> lsp_types::Range {
+        let start = self.position(range.start());
+        let end = self.position(range.end());
+        lsp_types::Range::new(start, end)
+    }
+
+    pub fn position(&self, offset: TextSize) -> lsp_types::Position {
+        let line_col = self.index.line_col(offset);
+        match self.encoding {
+            PositionEncoding::Utf8 => lsp_types::Position::new(line_col.line, line_col.col),
+            PositionEncoding::Wide(enc) => {
+                let line_col = self.index.to_wide(enc, line_col).unwrap();
+                lsp_types::Position::new(line_col.line, line_col.col)
+            }
+        }
+    }
+
+    pub fn try_range(&self, range: TextRange) -> Option<lsp_types::Range> {
+        let start = self.try_position(range.start())?;
+        let end = self.try_position(range.end())?;
+        lsp_types::Range::new(start, end).into()
+    }
+
+    pub fn try_position(&self, offset: TextSize) -> Option<lsp_types::Position> {
+        let line_col = self.index.try_line_col(offset)?;
+        match self.encoding {
+            PositionEncoding::Utf8 => lsp_types::Position::new(line_col.line, line_col.col).into(),
+            PositionEncoding::Wide(enc) => {
+                let line_col = self.index.to_wide(enc, line_col).unwrap();
+                lsp_types::Position::new(line_col.line, line_col.col).into()
+            }
+        }
+    }
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]