+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 ## Release Notes
 
+## [0.13.4]
+
+### Fixed
+
+- Fixed diagnostics not being converted to u_range.
+
 ## [0.13.3]
 
 ### Fixed
 
+- Completely rewrote the source mapping system for better preprocessor integration.
 - Fixed generic events not being included (see #411).
 - Fixed missing IntelliSense for some invalid syntax (see #406).
 - Fixed static expressions not being evaluated (see #412).