+++ editors/code/CHANGELOG.md
@@ -1,5 +1,17 @@
 ## Release Notes
 
+## [8.0.5]
+
+### Fixed
+
+- LSP version bump
+
+## [8.0.4]
+
+### Fixed
+
+- LSP version bump
+
 ## [8.0.3]
 
 ### Fixed